CR comments: moving credentials extraction logic to `ClientConfig`
