+++ client/src/main/java/io/pravega/client/ClientFactory.java
@@ -10,7 +10,6 @@
 package io.pravega.client;
 
 import com.google.common.annotations.Beta;
-import io.pravega.client.auth.CredentialsExtractor;
 import io.pravega.client.batch.BatchClient;
 import io.pravega.client.netty.impl.ConnectionFactoryImpl;
 import io.pravega.client.state.InitialUpdate;
@@ -74,7 +73,7 @@ public interface ClientFactory extends AutoCloseable {
      * @return Instance of ClientFactory implementation.
      */
     static ClientFactory withScope(String scope, ClientConfig config) {
-        config = CredentialsExtractor.extractCreds(config);
+        config = config.toBuilder().extractCredentials().build();
         val connectionFactory = new ConnectionFactoryImpl(config);
         return new ClientFactoryImpl(scope, new ControllerImpl(ControllerImplConfig.builder().clientConfig(config).build(),
                 connectionFactory.getInternalExecutor()), connectionFactory);
